Inline videos, Size toggle, and Code refactor
With this update, the plugin has achieved everything I believe it should've. I'm so glad I had to help to make it a reality.
Thank you greatly to NoSkillNoAndy for their invaluable help, slightly-less-novice-level skill, and late-night brainstorming sessions to make the functionality for this update possible, I never would've been able to do it without them. With that all being said, here are the additions.

- Inline Video Playback: At long last, videos that are uploaded as part of a mosaic will no longer open up into the carousel interface to play, they will play in chat like any video posted standalone.

- Size Toggle: The CSS for making images smaller (400px in width instead of the newer 500px) has been changed into a togglable setting in the plugin's new settings menu. It is enabled by default.

- Code Refactor: The plugin's code has been overhauled to be less messy and more refined.

Enjoy!