temporary patch for videos
With the new implementation (and smaller images enabled), videos that are part of a mosaic will re-crop themselves when played to be the same width as videos in the mosaic which are wider, cutting off the video and player controls. This small CSS fix prevents that by limiting the size videos can be, but they may become slightly bigger when played.