Restore Video Metadata
This update restores the original functionality where the top left-hand corner of uploaded video embeds would contain the name and size of the media and slide up when not in focus. This also tweaks the hover buttons' functionality in the embed's top right-hand corner to work the same way. It can be toggled off in the plugin's settings.

Special thanks to:
NoSkillPureAndy - For tying everything together and cleaning everything up
Arven (zrodevkaan) - For helping with the patch for the video player to get media info
Pix (11pixels) - For the original CSS styling & functionality demonstration (so I wasn't entirely reliant on old screenshots)
Rlego - For asking for this since I started working on the plugin